Extract FilterGroupActions
